Update and fix react-hot-loader
