+++ package.json
@@ -96,6 +96,7 @@
     "@types/react": "^0.0.0",
     "@types/react-dom": "^0.14.20",
     "@types/react-helmet": "^0.0.27",
+    "@types/react-hot-loader": "3.0.0-beta.6",
     "@types/react-redux": "^4.4.35",
     "@types/react-router": "^2.0.41",
     "@types/react-router-redux": "^4.0.36",
@@ -166,13 +167,13 @@
     "react": "^15.4.2",
     "react-dom": "^15.4.2",
     "react-helmet": "^3.3.0",
+    "react-hot-loader": "3.0.0-beta.6",
     "react-inline-css": "^2.3.0",
     "react-redux": "^5.0.1",
     "react-router": "^3.0.0",
     "react-router-redux": "^4.0.7",
     "react-tap-event-plugin": "^2.0.1",
     "react-transform-catch-errors": "^1.0.2",
-    "react-transform-hmr": "^1.0.4",
     "redbox-react": "^1.3.3",
     "redux": "^3.6.0",
     "redux-connect": "^5.0.0",
